Layouts

Sidebar layouts

Sidebar layouts split the page into one-third (for a control panel) and two-thirds (for the output). Each third is displayed in a card to visually separate it.

Matte.sidebar_layout — Function

Create a side bar layout for your app: a 1/3 width control panel and a 2/3 width main/output panel.

Inputs: leftpanel: The panel (either a main_panel or side_panel) to go on the left side of the layout. rightpanel: The panel (either a main_panel or side_panel) to go on the right side of the layout.

Matte.side_panel — Function

A narrow/side control (1/3 width) panel to be embedded in a sidebar_layout

Inputs: * content...: the content to fill the panel

Should only be used as part of a sidebar_layout

Matte.main_panel — Function

A main (2/3 width) panel to be embedded in a sidebar_layout

Inputs: * content...: the content to fill the panel

Should only be used as part of a sidebar_layout

Tab panel layout

Layout with pages/panels controlled by a tab bar.

Matte.tabs_layout — Function

Create a tab layout – app (or inset) with pages that can be navigated to using a series of tabs

Input: tabs...: A tuple of tab_panels to display in the layout vertical (optional, default false): Display the tabs as a vertical menu

Matte.tab_panel — Function

Define a panel for use in a tab layout

Inputs: title: Title for the tab (appears in the tab bar) content: Content for the panel

tab_panel should only be used as input to a tabs_layout

Custom layouts

Functions to create custom flexgrid layouts using Vuetify's built-in container layout.

Matte.custom_grid_layout — Function

Create a container to contain a custom layout.

Containers embed rows (which embed columns) to create layouts.

Inputs: contents...: The content to go in the container. Must be rows.

Matte.custom_grid_row — Function

Create a row for a custom layout

Inputs: content...: The content to go in the rows. Must be columns.

Rows wrap columns.

Matte.custom_grid_column — Function

Create a layout column that covers width/12 of the row

Inputs: * content...: The content to go in the column * width: Width of the column. Full width = 12 . (Default 12)

Columns must be contained in rows (custom_grid_row)

Matte.custom_card — Function

Create a 'card' to contain content. Cards are boxes with drop shadows.

For use in custom_grid_layouts, to define sections of the page inside a column

Inputs: content...: The content to be contained inside the card