Running and creating apps

Run

Matte.run_app — Function

run_app(app; async = true)

Theapp module that defines your app must have a function called ui that defines the UI. The functions in the module (which don't need to be exported) define the logic for your app. The name of the function is the name of the outputs in your UI, and the inputs to those functions correspond to the ids of inputs in the UI.

See the getting started guide for help on how to define your app.

Matte.stop_app — Function

stop_app()

Stops a running server that was started with async=true

Helpers to create apps

Matte.new_matte_app — Function

new_matte_app(path; force = false, activate = true)

Create a new skeleton Matte app in the folder path. The folder should be empty. Matte will create the final folder if it doesn't exist, but will fail if the rest of the path doesn't exist.

path should be a folder that doesn't exist. If it doesnt, you can use force = true to overwrite. activate controls whether the helper will activate the project for the newly created app.

Matte.matte_example — Function

matte_example(example, path; force = false, activate = true)

Create a new Matte app, based on the example specified by example. See the manual for a list of examples that you can use here.

path should be a folder that doesn't exist. If it doesnt, you can use force = true to overwrite. activate controls whether the helper will activate the project for the newly created app.