Running and creating apps

Run

Matte.run_app — Function

Run a Matte app

Inputs: app: The module that defines your app. async (optional, default false): Start the server in async mode (returns a repl after starting the server).

The module that defines your app must have a function called ui that defines the UI. The functions in the module (which don't need to be exported) define the logic for your app. The name of the function is the name of the outputs in your UI, and the inputs to those functions correspond to the ids of inputs in the UI.

See the getting started guide for help on how to define your app.

Helpers to create apps

Matte.new_matte_app — Function

new_matte_app(path, force = false)

Create a new skeleton Matte app in the folder path. The folder should be empty. Matte will create the final folder if it doesn't exist, but will fail if the rest of the path doesn't exist.

path should be a folder that doesn't exist. If it doesnt, you can use force = true to overwrite

Matte.matte_example — Function

matte_example(example, path, force = false)

Create a new Matte app, based on the example specified by example. See the manual for a list of examples that you can use here.

path should be a folder that doesn't exist. If it doesnt, you can use force = true to overwrite