Controlling UI look and style

Basic text formatting

Matte.p — Function

p(contents)

Wrap some text in a paragraph tag. Helps with layout of raw text.

Matte.h1 — Function

h1(contents...)

Show contents as a level 1 heading

Matte.h2 — Function

h2(contents...)

Show contents as a level 2 heading

Matte.h3 — Function

h3(contents...)

Show contents as a level 3 heading

Matte.br — Function

br()

Insert a line break (html <br>) into your UI

Notifying users of events

Matte.dialog — Function

dialog(id, title, content, width = 500)

Display a dialog box (which pops over the rest of the content, forcing user acknowledgement). Visibility is controlled by id. A common use case is to define this output function to have sole input from a button, and simply return whether the button has been clicked.

title and content can be dynamically set.

If you need to include multiple elements inside content, wrap them as a tuple.

Note

For best results dialogs should be placed at the end of your UI, after the layout. Placing them inside a layout can cause them to be hidden.