Input elements

Matte.slider — Function

Input element that allows users to select among a set of integers using a slider

Inputs: id: Unique id for the input. This id is used as the variable name for the value of in your server-side functions label: Some text to describe what the slider does min: Minimum value of the slider max: Maximum value of the slider default (optional): The starting value for the slider. Defaults to the mid point

Matte.text_input — Function

Free text field for string input by users

Inputs: id: Unique id for the input. This id is used as the variable name for the value of in your server-side functions. label: Some text to describe what the text input is for default: The starting value for the input when the app loads (default empty string)

Note that a text input always returns a String to the server. If you need users to be able to freely enter a number, use number_input.

Matte.number_input — Function

Form input that only accepts numbers

Inputs: id: Unique id for the input. This id is used as the variable name for the value of in your server-side functions. label: Some text to describe what the number input is for default: The starting value for the input when the app loads (default 0)

Note that number_input always returns a float to the server. If you need a string use text_input. If you need integers, you can use a slider or round the result.