Layouts

Sidebar layouts

Sidebar layouts split the page into one-third (for a control panel) and two-thirds (for the output). Each third is displayed in a card to visually separate it.

Matte.sidebar_layout — Function

sidebar_layout(leftpanel, rightpanel)

Create a side bar layout for your app: a 1/3 width control panel and a 2/3 width main/output panel.

Matte.side_panel — Function

side_panel(content...)

A narrow/side control (1/3 width) panel to be embedded in a sidebar_layout

Should only be used as part of a sidebar_layout

Matte.main_panel — Function

 main_panel(content...)

A main (2/3 width) panel to be embedded in a sidebar_layout

Should only be used as part of a sidebar_layout

Tab panel layout

Layout with pages/panels controlled by a tab bar.

Matte.tabs_layout — Function

tabs_layout(tabs...; vertical = false)

Create a tab layout – app (or inset) with pages that can be navigated to using a series of tabs

Matte.tab_panel — Function

tab_panel(title, content)

Define a panel for use in a tab layout

tab_panel should only be used as input to a tabs_layout

Custom layouts

Functions to create custom flexgrid layouts using Vuetify's built-in container layout.

Matte.custom_grid_layout — Function

custom_grid_layout(content...)

Create a container to contain a custom layout.

Containers embed rows (which embed columns) to create layouts.

Matte.custom_grid_row — Function

custom_grid_row(content...)

Create a row for a custom layout

Rows wrap columns.

Matte.custom_grid_column — Function

custom_grid_column(content...; width = 12)

Create a layout column that covers width/12 of the row

Columns must be contained in rows (custom_grid_row)

Matte.custom_card — Function

custom_card(content...)

Create a 'card' to contain content. Cards are boxes with drop shadows.

For use in custom_grid_layouts, to define sections of the page inside a column